requirements.txt
fastapi>=0.109.0,<0.115.0
uvicorn[standard]>=0.27.0,<0.32.0
sqlalchemy[asyncio]>=2.0.0,<2.1.0
asyncpg>=0.29.0,<0.30.0
geoalchemy2>=0.14.0,<0.15.0
pydantic[email]>=2.5.0,<2.10.0
pydantic-settings>=2.1.0,<2.6.0
alembic>=1.13.0,<1.14.0
passlib[argon2]>=1.7.4,<1.8.0
python-ulid>=2.0.0,<3.0.0
shapely>=2.0.0,<3.0.0
